# backend/main.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import List, Dict

app = FastAPI()

class IngestRequest(BaseModel):
    content_path: str

class ChatRequest(BaseModel):
    query: str
    selected_text: str = None

@app.post("/ingest")
async def ingest_content(request: IngestRequest):
    """
    Ingests content from the specified path into the vector database.
    """
    try:
        # Placeholder for ingestion logic
        # In a real implementation, this would call backend/rag/ingest.py
        print(f"Ingesting content from: {request.content_path}")
        return {"status": "success", "message": f"Content from {request.content_path} ingested."}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/chat")
async def chat_with_rag(request: ChatRequest):
    """
    Handles RAG chatbot queries.
    """
    try:
        # Placeholder for chat logic
        # In a real implementation, this would call backend/rag/retrieve.py and backend/rag/chain.py
        response_message = f"You asked: '{request.query}'. "
        if request.selected_text:
            response_message += f"Context from selected text: '{request.selected_text}'. "
        response_message += "This is a placeholder response."

        return {"status": "success", "response": response_message, "citations": []}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/")
async def read_root():
    return {"message": "Physical AI & Humanoid Robotics RAG Backend is running!"}