# specs/001-physical-ai-book/contracts/api-contracts.yaml
openapi: 3.0.0
info:
  title: Physical AI & Humanoid Robotics RAG Backend API
  version: 1.0.0
  description: API for ingesting textbook content and interacting with the RAG chatbot, with future extensions for user authentication, personalization, and translation.

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /ingest:
    post:
      summary: Ingest Content
      description: Triggers the ingestion pipeline to load MDX content into the vector database. This operation is typically performed automatically via CI/CD (e.g., GitHub Actions).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Content ingestion initiated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '500':
          description: Internal server error during ingestion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chat:
    post:
      summary: Chat with RAG Chatbot
      description: Sends a query to the RAG chatbot to get answers based on the ingested textbook content. Supports optional selected text for contextualized answers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response from the RAG chatbot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '500':
          description: Internal server error during chat processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    IngestRequest:
      type: object
      required:
        - content_path
      properties:
        content_path:
          type: string
          description: The path to the directory containing the MDX content to ingest (e.g., "./docs").
          example: "../../docs"

    IngestResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          enum: ["success", "error"]
          example: "success"
        message:
          type: string
          example: "Content from ../../docs ingested."

    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's question to the RAG chatbot.
          example: "What are the core concepts of ROS 2?"
        selected_text:
          type: string
          nullable: true
          description: Optional. A highlighted text fragment from the textbook to provide additional context for the query.
          example: "ROS 2 Nodes, Topics, and Services"

    Citation:
      type: object
      required:
        - source
        - page_content
      properties:
        source:
          type: string
          description: The source file or URL of the cited content.
          example: "module-01-ros2/01-ros2-concepts.mdx"
        page_content:
          type: string
          description: The exact snippet of content from the source that supports the answer.
          example: "ROS 2 (Robot Operating System 2) is a flexible framework for writing robot software."

    ChatResponse:
      type: object
      required:
        - status
        - response
        - citations
      properties:
        status:
          type: string
          enum: ["success", "error"]
          example: "success"
        response:
          type: string
          description: The answer generated by the RAG chatbot.
          example: "ROS 2 is a flexible framework for writing robot software, offering distributed architecture and QoS."
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
          description: A list of citations from the textbook content that support the answer.

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "An internal server error occurred."

  # Placeholder schemas for future phases
    UserProfile:
      type: object
      properties:
        user_id:
          type: string
        python_experience:
          type: string
        ros_experience:
          type: string
        gpu_available:
          type: boolean
        jetson_access:
          type: boolean
        robot_access:
          type: boolean

    PersonalizationRequest:
      type: object
      required:
        - chapter_id
      properties:
        chapter_id:
          type: string

    PersonalizationResponse:
      type: object
      required:
        - status
        - personalized_content
      properties:
        status:
          type: string
        personalized_content:
          type: string

    TranslationRequest:
      type: object
      required:
        - chapter_id
        - target_language
      properties:
        chapter_id:
          type: string
        target_language:
          type: string

    TranslationResponse:
      type: object
      required:
        - status
        - translated_content
      properties:
        status:
          type: string
        translated_content:
          type: string